+++ tests/test_facebook_field_selection.py
@@ -25,6 +25,7 @@ class FacebookFieldSelection(FacebookBaseTest):  # TODO use base.py, determine i
             'ads_insights_platform_and_device',
             'ads_insights_region',
             'ads_insights_dma',
+            "ads_insights_hourly_advertiser",
             #'leads',
         }
 
@@ -41,6 +42,7 @@ class FacebookFieldSelection(FacebookBaseTest):  # TODO use base.py, determine i
             "ads_insights_platform_and_device",
             "ads_insights_region",
             "ads_insights_dma",
+            "ads_insights_hourly_advertiser",
             #"leads",
         }
 
@@ -53,10 +55,11 @@ class FacebookFieldSelection(FacebookBaseTest):  # TODO use base.py, determine i
             "campaigns" :                       {"id"},
             "ads_insights" :                    {"campaign_id", "adset_id", "ad_id", "date_start"},
             "ads_insights_age_and_gender" :     {"campaign_id", "adset_id", "ad_id", "date_start", "age", "gender"},
-            "ads_insights_country" :            {"campaign_id", "adset_id", "ad_id", "date_start"},
+            "ads_insights_country" :            {"campaign_id", "adset_id", "ad_id", "date_start", "country"},
             "ads_insights_platform_and_device": {"campaign_id", "adset_id", "ad_id", "date_start", "publisher_platform", "platform_position", "impression_device"},
             "ads_insights_region" :             {"campaign_id", "adset_id", "ad_id", "date_start"},
             "ads_insights_dma" :                {"campaign_id", "adset_id", "ad_id", "date_start"},
+            "ads_insights_hourly_advertiser":   {"campaign_id", "adset_id", "ad_id", "date_start", "hourly_stats_aggregated_by_advertiser_time_zone"},
             #"leads" :                           {"id"},
         }